split generic parser and utils
